# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "API Super Tienda",
    "contact": {
      "name": "Super    Tienda",
      "url": "https://www.superTienda.com",
      "email": "superTienda@gmail.com"
    }
  },
  "components": {
    "schemas": {
...
      "BusinessProduct": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "businessId": {
            "type": "integer"
          },
          "productId": {
            "type": "integer"
          },
          "customPrice": {
            "type": "number"
          },
          "currentStock": {
            "type": "integer",
            "default": 0
          },
          "reservedStock": {
            "type": "integer",
            "default": 0
          },
          "availableStock": {
            "type": "integer",
            "default": 0
          },
          "lastRestock": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "business": {
            "$ref": "#/definitions/Business"
          },
          "product": {
            "$ref": "#/definitions/Product"
          }
        },
        "required": [
          "businessId",
          "productId",
          "customPrice"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/business-products": {
      "get": {
        "summary": "Listar todos los productos asignados a negocios",
        "tags": [
          "BusinessProducts"
        ],
        "responses": {
          "200": {
            "description": "Lista de asignaciones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BusinessProduct"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear una nueva asignación de producto a negocio",
        "tags": [
          "BusinessProducts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "businessId",
                  "productId"
                ],
                "properties": {
                  "businessId": {
                    "type": "integer"
                  },
                  "productId": {
                    "type": "integer"
                  },
                  "customPrice": {
                    "type": "number",
                    "format": "decimal"
                  },
                  "actualStock": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Asignación creada"
          },
          "400": {
            "description": "Asignación duplicada"
          }
        }
      }
    },
    "/api/business-products/{id}": {
      "get": {
        "summary": "Obtener asignación de producto por ID",
        "tags": [
          "BusinessProducts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asignación encontrada"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar asignación de producto a negocio",
        "tags": [
          "BusinessProducts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customPrice": {
                    "type": "number"
                  },
                  "actualStock": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Asignación actualizada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar asignación de producto",
        "tags": [
          "BusinessProducts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado exitosamente"
          }
        }
      }
    },
...
    "/api/cart-items": {
      "get": {
        "summary": "Listar todos los ítems de carritos",
        "tags": [
          "CartItems"
        ],
        "responses": {
          "200": {
            "description": "Lista de ítems"
          }
        }
      },
      "post": {
        "summary": "Agregar producto al carrito",
        "tags": [
          "CartItems"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "cart_id",
                  "product_id",
                  "quantity",
                  "price_at_time"
                ],
                "properties": {
                  "cart_id": {
                    "type": "integer"
                  },
                  "product_id": {
                    "type": "integer"
                  },
                  "quantity": {
                    "type": "integer"
                  },
                  "price_at_time": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ítem agregado"
          }
        }
      }
    },
    "/api/cart-items/{id}": {
      "get": {
        "summary": "Obtener ítem de carrito por ID",
        "tags": [
          "CartItems"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ítem encontrado"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar producto en carrito",
        "tags": [
          "CartItems"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "integer"
                  },
                  "price_at_time": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ítem actualizado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar ítem del carrito",
        "tags": [
          "CartItems"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Ítem eliminado"
          }
        }
      }
    },
...
  },
...
}